Convert a number of ints to size_t/FX_FILESIZE in fxge
These are the straightforward ones.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/86930
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ core/fpdfapi/page/cpdf_dib.cpp
@@ -1203,7 +1203,7 @@ bool CPDF_DIB::SkipToScanline(int line, PauseIndicatorIface* pPause) const {
   return m_pDecoder && m_pDecoder->SkipToScanline(line, pPause);
 }
 
-uint32_t CPDF_DIB::GetEstimatedImageMemoryBurden() const {
+size_t CPDF_DIB::GetEstimatedImageMemoryBurden() const {
   return m_pCachedBitmap ? m_pCachedBitmap->GetEstimatedImageMemoryBurden() : 0;
 }
