+++ core/fxge/cfx_fontmapper.cpp
@@ -195,7 +195,7 @@ void GetFontFamily(uint32_t nStyle, ByteString* fontName) {
   }
 }
 
-ByteString ParseStyle(const char* pStyle, int iLen, int iIndex) {
+ByteString ParseStyle(const char* pStyle, size_t iLen, size_t iIndex) {
   std::ostringstream buf;
   if (!iLen || iLen <= iIndex)
     return ByteString(buf);
@@ -472,7 +472,7 @@ RetainPtr<CFX_Face> CFX_FontMapper::FindSubstFont(const ByteString& name,
       }
     }
     if (!bHasHyphen) {
-      int nLen = family.GetLength();
+      size_t nLen = family.GetLength();
       bool hasStyleType;
       uint32_t styleType;
       size_t len;
@@ -490,9 +490,9 @@ RetainPtr<CFX_Face> CFX_FontMapper::FindSubstFont(const ByteString& name,
     weight = FXFONT_FW_BOLD;
 
   if (!style.IsEmpty()) {
-    int nLen = style.GetLength();
+    size_t nLen = style.GetLength();
     const char* pStyle = style.c_str();
-    int i = 0;
+    size_t i = 0;
     bool bFirstItem = true;
     ByteString buf;
     while (i < nLen) {
